Basic Activities and Fragments. Some Minor Elements of UI. Basic Elements Necessary to Implement Search
