+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,20 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/ListViewFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.example.shoppinglist.FirstFragment"
+        android:id="@+id/EditListFragment"
+        android:name="com.example.shoppinglist.EditListFragment"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_edit_list" />
     <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.shoppinglist.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
+        android:id="@+id/ListViewFragment"
+        android:name="com.example.shoppinglist.ListViewFragment"
+        android:label="@string/first_fragment_label"
+        tools:layout="@layout/fragment_list_view">
 
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
+
     </fragment>
+
 </navigation>
